Replace bower-installer with bower - separate external dependancy with internal resource - rename static to resources - rename templates to partials
