+++ src/web/partials/admin/index.html
