+++ src/web/partials/character.html
@@ -11,7 +11,8 @@
         <button class="btn btn-success" ng-click="newPlayer()"><span class="glyphicon glyphicon-plus"></span> <span class="glyphicon glyphicon-user"></span> Nouveau joueur</button>
       </div>
       <div ng-show="new_player">
-        <button class="btn btn-success" ng-click="submitCharacterData()"><span class="glyphicon glyphicon-floppy-save"></span> <span class="glyphicon glyphicon-user"></span> Enregistrer nouveau joueur</button>
+        <button class="btn btn-success" ng-click="submitCharacterData()"><span class="glyphicon glyphicon-floppy-save"></span> <span class="glyphicon glyphicon-user"></span> Enregistrer nouveau joueur
+        </button>
       </div>
       <div ng-show="new_player">
         <button class="btn btn-danger" ng-click="discardPlayer()"><span class="glyphicon glyphicon-remove"></span> <span class="glyphicon glyphicon-user"></span> Annuler nouveau joueur</button>
@@ -64,6 +65,15 @@
         </div>
       </div>
 
+      <div class="form-group row">
+        <label class="col-sm-2 form-control-label">Fraction</label>
+        <div class="col-sm-8" class="btn-group" data-toggle="buttons">
+          <label class="btn btn-primary active" ng-repeat="opt in rule.user.faction">
+            <input type="radio" name="faction_options" autocomplete="off" checked>Option1
+          </label>
+        </div>
+      </div>
+
       <div class="print-hide form-group row">
         <div class="col-sm-offset-2 col-sm-8">
           <button type="submit" ng-click="submitCharacterData()" class="btn btn-success"><span class="glyphicon glyphicon-floppy-save"></span> Enregistrer</button>
@@ -96,7 +106,7 @@
             <button type="submit" ng-click="submitCharacterData()" class="btn btn-success"><span class="glyphicon glyphicon-floppy-save"></span> Enregistrer</button>
             <input type="button" value="Imprimer" class="btn btn-primary" ng-click="printCharacterSheet()"/>
             <button type="submit" ng-click="discardPlayer()" class="btn btn-danger"><span class="glyphicon glyphicon-remove"></span> Annuler</button>
-            <button type="submit" ng-click="deleteCharacter()" class="btn btn-danger"  ng-hide="new_character || new_player"><span class="glyphicon glyphicon-floppy-remove"></span> Supprimer</button>
+            <button type="submit" ng-click="deleteCharacter()" class="btn btn-danger" ng-hide="new_character || new_player"><span class="glyphicon glyphicon-floppy-remove"></span> Supprimer</button>
           </div>
         </div>
