+++ src/web/partials/login.html
@@ -3,8 +3,8 @@
 {% block content %}
 
 <!--Source http://bootsnipp.com/snippets/featured/responsive-login-with-social-buttons -->
-<link href="{{ static_url('local/css/login.css') }}" rel="stylesheet">
-<link href="{{ static_url('local/css/login_dark.css') }}" rel="stylesheet">
+<link href="{{ static_url('resources/css/login.css') }}" rel="stylesheet">
+<link href="{{ static_url('resources/css/login_dark.css') }}" rel="stylesheet">
 
 <div class="container" ng-controller="login_ctrl">
   <!--Sign in-->
