+++ src/web/partials/news.html
@@ -4,7 +4,7 @@
 
 <div ng-controller="news_ctrl">
   {% if enable_facebook_feed %}
-  <script src="{{ static_url('local/js/facebook.js') }}"></script>
+  <script src="{{ static_url('resources/js/facebook.js') }}"></script>
 
   <h1>Page facebook!</h1>
   <!-- begin facebook page traitre-lame -->
