+++ src/web/py_class/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
