+++ src/web/resources/icon/favicon.ico
