+++ src/web/resources/js/tool.js
